requirements.txt
python-dotenv
langchain_google_genai
langchain
langchain-core